feat: extract buildPrompt and snapshot util
